fix: move coef to Repository struct
