+++ chain/metrics.go
@@ -3,10 +3,10 @@
 // Distributed under the GNU Lesser General Public License v3.0 software license, see the accompanying
 // file LICENSE or <https://www.gnu.org/licenses/lgpl-3.0.html>
 
-package consensus
+package chain
 
 import "github.com/vechain/thor/v2/metrics"
 
 var (
-	metricCoefBucket = metrics.LazyLoadHistogram("consensus_coef_histogram", metrics.BucketCoefficients)
+	metricCoefBucket = metrics.LazyLoadHistogram("chain_coef_histogram", metrics.BucketCoefficients)
 )
